+++ src/jsMain/kotlin/logic/Tile.kt
@@ -7,7 +7,7 @@ import react.Props
 import react.dom.html.ReactHTML
 
 interface Tile {
-    fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox>
+    fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox>
     fun ChildrenBuilder.render(colors: Colors)
 }
 
@@ -19,7 +19,7 @@ external interface TileProps : Props {
 val drawTile = FC<TileProps> { props -> props.tile.run { render(props.colors) } }
 
 open class Empty : Tile {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors to fox.go()
 
     override fun ChildrenBuilder.render(colors: Colors) {}
@@ -30,7 +30,7 @@ object Start : Empty() {
 }
 
 class End : Tile {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors to fox
 
     override fun ChildrenBuilder.render(colors: Colors) = +"#"
@@ -49,9 +49,9 @@ abstract class ColoredTile(protected var colorIndex: Int?) : Tile {
 }
 
 class Arrow(colorIndex: Int?, private val direction: Direction) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> {
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> {
         return black to colors to (
-                if ((colorIndex?.let { colors[colorIndex!!].value } ?: 0) == 0) fox.go(direction) else fox.go()
+                if ((colorIndex?.let { colors[colorIndex!!].value } ?: 0L) == 0L) fox.go(direction) else fox.go()
                 )
     }
 
@@ -61,7 +61,7 @@ class Arrow(colorIndex: Int?, private val direction: Direction) : ColoredTile(co
 }
 
 class Inc(colorIndex: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors.inc(colorIndex!!) to fox.go()
 
     override fun ChildrenBuilder.coloredRender() {
@@ -70,7 +70,7 @@ class Inc(colorIndex: Int) : ColoredTile(colorIndex) {
 }
 
 class Dec(colorIndex: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors.dec(colorIndex!!) to fox.go()
 
     override fun ChildrenBuilder.coloredRender() {
@@ -78,8 +78,8 @@ class Dec(colorIndex: Int) : ColoredTile(colorIndex) {
     }
 }
 
-class Constant(colorIndex: Int, private val value: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+class Constant(colorIndex: Int, private val value: Long) : ColoredTile(colorIndex) {
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors.assign(colorIndex!!, value) to fox.go()
 
     override fun ChildrenBuilder.coloredRender() {
@@ -88,7 +88,7 @@ class Constant(colorIndex: Int, private val value: Int) : ColoredTile(colorIndex
 }
 
 class Store(colorIndex: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         colors[colorIndex!!].value to colors to fox.go()
 
     override fun ChildrenBuilder.coloredRender() {
@@ -97,7 +97,7 @@ class Store(colorIndex: Int) : ColoredTile(colorIndex) {
 }
 
 class Plus(colorIndex: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors.plus(colorIndex!!, black) to fox.go()
 
     override fun ChildrenBuilder.coloredRender() {
@@ -106,7 +106,7 @@ class Plus(colorIndex: Int) : ColoredTile(colorIndex) {
 }
 
 class Minus(colorIndex: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors.minus(colorIndex!!, black) to fox.go()
 
     override fun ChildrenBuilder.coloredRender() {
@@ -115,7 +115,7 @@ class Minus(colorIndex: Int) : ColoredTile(colorIndex) {
 }
 
 class Times(colorIndex: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors.times(colorIndex!!, black) to fox.go()
 
     override fun ChildrenBuilder.coloredRender() {
@@ -124,7 +124,7 @@ class Times(colorIndex: Int) : ColoredTile(colorIndex) {
 }
 
 class Div(colorIndex: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors.div(colorIndex!!, black) to fox.go()
 
     override fun ChildrenBuilder.coloredRender() {
@@ -133,7 +133,7 @@ class Div(colorIndex: Int) : ColoredTile(colorIndex) {
 }
 
 class Rem(colorIndex: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors.rem(colorIndex!!, black) to fox.go()
 
     override fun ChildrenBuilder.coloredRender() {
@@ -142,9 +142,9 @@ class Rem(colorIndex: Int) : ColoredTile(colorIndex) {
 }
 
 class Input(colorIndex: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> {
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> {
         val (nextFox, char) = fox.go(input)
-        return black to colors.assign(colorIndex!!, char.code) to nextFox
+        return black to colors.assign(colorIndex!!, char.code.toLong()) to nextFox
     }
 
     override fun ChildrenBuilder.coloredRender() {
@@ -153,8 +153,8 @@ class Input(colorIndex: Int) : ColoredTile(colorIndex) {
 }
 
 class Output(colorIndex: Int) : ColoredTile(colorIndex) {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
-        black to colors to fox.go(colors[colorIndex!!].value.toChar())
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
+        black to colors to fox.go(colors[colorIndex!!].value.toInt().toChar())
 
     override fun ChildrenBuilder.coloredRender() {
         +"o"
@@ -162,7 +162,7 @@ class Output(colorIndex: Int) : ColoredTile(colorIndex) {
 }
 
 class Jump(private val position: Position) : Tile {
-    override fun go(black: Int, colors: Colors, fox: Fox, input: String): Pair<Pair<Int, Colors>, Fox> =
+    override fun go(black: Long, colors: Colors, fox: Fox, input: String): Pair<Pair<Long, Colors>, Fox> =
         black to colors to fox.go(position)
 
     override fun ChildrenBuilder.render(colors: Colors) {
